MISUV-7256:   * - extract trait from FeatureSwitchServiceImpl and other relevant refactoring;
