+++ app/services/admin/FeatureSwitchService.scala
@@ -16,64 +16,18 @@
 
 package services.admin
 
-import config.featureswitch.FeatureSwitching
-import config.{AgentItvcErrorHandler, FrontendAppConfig, ItvcErrorHandler}
-import controllers.agent.predicates.ClientConfirmedController
+import com.google.inject.ImplementedBy
 import models.admin.{FeatureSwitch, FeatureSwitchName}
-import play.api.Logger
-import play.api.mvc.MessagesControllerComponents
-import repositories.admin.FeatureSwitchRepository
-import uk.gov.hmrc.auth.core.AuthorisedFunctions
 
-import javax.inject.{Inject, Singleton}
-import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.Future
 
-@Singleton
-class FeatureSwitchService @Inject()(
-                                      val featureSwitchRepository: FeatureSwitchRepository,
-                                      val authorisedFunctions: AuthorisedFunctions,
-                                      val appConfig: FrontendAppConfig)
-                                    (implicit val ec: ExecutionContext,
-                                     mcc: MessagesControllerComponents,
-                                     implicit val itvcErrorHandler: ItvcErrorHandler,
-                                     implicit val itvcErrorHandlerAgent: AgentItvcErrorHandler) extends ClientConfirmedController with FeatureSwitching {
+@ImplementedBy(classOf[FeatureSwitchServiceImpl])
+trait FeatureSwitchService {
+  def get(featureSwitchName: FeatureSwitchName): Future[FeatureSwitch]
 
-  def get(featureSwitchName: FeatureSwitchName): Future[FeatureSwitch] =
-    featureSwitchRepository
-      .getFeatureSwitch(featureSwitchName)
-      .map(_.getOrElse(FeatureSwitch(featureSwitchName, false)))
+  def getAll: Future[List[FeatureSwitch]]
 
+  def set(featureSwitchName: FeatureSwitchName, enabled: Boolean): Future[Boolean]
 
-  def getAll: Future[List[FeatureSwitch]] = {
-
-    Logger("application").info(s"reading FSS - ${appConfig.readFeatureSwitchesFromMongo}")
-    if (appConfig.readFeatureSwitchesFromMongo) {
-      // TODO: do we need to apply fallback in case can not connect to mongoDb?
-      featureSwitchRepository.getFeatureSwitches.map { mongoSwitches =>
-        Logger("application").info(s"reading FSS: ${mongoSwitches}")
-        FeatureSwitchName.allFeatureSwitches
-          .foldLeft(mongoSwitches) { (featureSwitches, missingSwitch) =>
-            if (featureSwitches.map(_.name).contains(missingSwitch))
-              featureSwitches
-            else
-              FeatureSwitch(missingSwitch, false) :: featureSwitches
-          }
-          .reverse
-      }
-    } else {
-      Future.successful(getFSList)
-    }
-  }
-
-  def set(featureSwitchName: FeatureSwitchName, enabled: Boolean): Future[Boolean] =
-    if (appConfig.readFeatureSwitchesFromMongo) {
-      featureSwitchRepository.setFeatureSwitch(featureSwitchName, enabled)
-    } else {
-      setFS(featureSwitchName, enabled)
-      Future.successful(true)
-    }
-
-  def setAll(featureSwitches: Map[FeatureSwitchName, Boolean]): Future[Unit] =
-    featureSwitchRepository.setFeatureSwitches(featureSwitches)
-
+  def setAll(featureSwitches: Map[FeatureSwitchName, Boolean]): Future[Unit]
 }
