+++ app/testOnly/controllers/FeatureSwitchController.scala
@@ -22,7 +22,7 @@ import models.admin.FeatureSwitchName
 import models.admin.FeatureSwitchName.allFeatureSwitches
 import play.api.i18n.I18nSupport
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
-import services.admin.FeatureSwitchService
+import services.admin.{FeatureSwitchService}
 import testOnly.views.html.FeatureSwitchView
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendController
