+++ app/testOnly/controllers/FeatureSwitchesAdminController.scala
@@ -19,7 +19,7 @@ package controllers.admin
 import models.admin.{FeatureSwitch, FeatureSwitchName}
 import play.api.libs.json.{JsBoolean, Json}
 import play.api.mvc.{AbstractController, Action, AnyContent, ControllerComponents}
-import services.admin.FeatureSwitchService
+import services.admin.{FeatureSwitchService}
 
 import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
