+++ it/test/helpers/agent/ComponentSpecBase.scala
@@ -88,7 +88,7 @@ trait ComponentSpecBase extends TestSuite with CustomMatchers
     "auditing.enabled" -> "true",
     "microservice.services.contact-frontend.host" -> mockHost,
     "microservice.services.contact-frontend.port" -> mockPort,
-    "feature-switches.read-from-mongo" -> "false"
+    "feature-switches.read-from-mongo" -> "true"
   )
 
   val userDetailsUrl = "/user-details/id/5397272a3d00003d002f3ca9"
