+++ test/controllers/ForecastTaxCalcSummaryControllerSpec.scala
@@ -28,7 +28,7 @@ import play.api.http.Status.{INTERNAL_SERVER_ERROR, NOT_FOUND, OK}
 import play.api.mvc.MessagesControllerComponents
 import play.api.test.Helpers.{charset, contentType, defaultAwaitTimeout, status}
 import services.IncomeSourceDetailsService
-import services.admin.FeatureSwitchService
+import services.admin.FeatureSwitchServiceImpl
 import testConstants.BaseTestConstants.{testAgentAuthRetrievalSuccess, testMtditid, testMtditidAgent, testTaxYear}
 import testUtils.TestSupport
 import views.html.ForecastTaxCalcSummary
@@ -47,7 +47,7 @@ class ForecastTaxCalcSummaryControllerSpec extends TestSupport with MockCalculat
     app.injector.instanceOf[ItvcErrorHandler],
     app.injector.instanceOf[IncomeSourceDetailsService],
     mockAuthService,
-    app.injector.instanceOf[FeatureSwitchService]
+    app.injector.instanceOf[FeatureSwitchServiceImpl]
   )(
     ec,
     languageUtils,
